areas cards link to active template and area
