+++ src/components/area-card/AreaCardContainer.js
@@ -1,9 +1,12 @@
 import { connect } from 'react-redux';
 import AreaCard from './AreaCard';
 
-const mapStateToProps = ({ areas }, { id }) => {
+const mapStateToProps = ({ areas, templates }, { id }) => {
   const area = areas.data[id] && areas.data[id].attributes;
-  return { area: { ...area, id } };
+  return { 
+    area: { ...area, id },
+    templates
+  };
 };
 
 export default connect(mapStateToProps, null)(AreaCard);
