+++ src/helpers/login.js
@@ -1,0 +1,13 @@
+  import querystring from 'query-string';
+  
+  const trimQueryParams = (props) => {
+    const { location, history } = props;
+    const search = location.search || '';
+    const queryParams = querystring.parse(search);
+    if (queryParams.token) {
+      const newSearch = querystring.stringify({ ...queryParams, token: undefined });
+      history.replace('/areas', { search: newSearch });
+    }
+  }
+
+  export { trimQueryParams };
